feat: Remove other vehicles for now, code cleanup & fix dependencies
