+++ Assets/Scripts/RidersGame/Analytics/InputProcessor.cs
@@ -1,11 +1,10 @@
+using RidersRuntime.DebugTools;
+using RidersRuntime.Input;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using UnityEngine;
 using UnityEngine.InputSystem;
-using RidersRuntime.Data;
-using RidersRuntime.Input;
-using RidersRuntime.DebugTools;
 
 namespace RidersRuntime.Analytics
 {
