--- Assets/Scripts/RidersGame/Enums.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 405ffb8b9493a55478fa30203850a3c4
