+++ Assets/Scripts/RidersGame/Inputs/BaseInput.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-using RidersRuntime.Data;
 
 namespace RidersRuntime.Input
 {
