+++ Assets/Scripts/RidersGame/System/DebugFlags.cs
@@ -1,7 +1,8 @@
+using RidersRuntime.Data;
+using RidersRuntime.Input;
 using System.IO;
 using System.Reflection;
 using UnityEngine;
-using RidersRuntime.Data;
 
 namespace RidersRuntime.DebugTools
 {
