+++ Assets/Scripts/RidersGame/UI/BaseVehicleDebug.cs
@@ -1,5 +1,5 @@
-using UnityEngine;
 using TMPro;
+using UnityEngine;
 
 namespace RidersRuntime.UI
 {
