--- Assets/Scripts/RidersGame/Values.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c43cdea51e544054d85f278053ddac58
