--- Assets/Scripts/RidersGame/VehicleSystems/BikeVehicle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4419568429669e446a6be19e3443546e
