--- Assets/Scripts/RidersGame/VehicleSystems/BladesVehicle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4ef3ec4f163068a49b6c3302714fefa8
