+++ Assets/Scripts/RidersGame/VehicleSystems/State Machine/Concrete States/MovementAirState.cs
@@ -1,5 +1,3 @@
-using UnityEngine;
-
 namespace RidersRuntime.VehicleSystem
 {
     public class MovementAirState : MovementState
