+++ Assets/Scripts/RidersGame/VehicleSystems/State Machine/Concrete States/MovementGrindState.cs
@@ -1,5 +1,3 @@
-using UnityEngine;
-
 namespace RidersRuntime.VehicleSystem
 {
     public class MovementGrindState : MovementState
