+++ Assets/Scripts/RidersGame/VehicleSystems/State Machine/Concrete States/MovementGroundState.cs
@@ -1,5 +1,3 @@
-using UnityEngine;
-
 namespace RidersRuntime.VehicleSystem
 {
     public class MovementGroundState : MovementState
