+++ Assets/Scripts/RidersGame/VehicleSystems/State Machine/MovementState.cs
@@ -1,5 +1,3 @@
-using UnityEngine;
-
 namespace RidersRuntime.VehicleSystem
 {
     public class MovementState
