+++ Assets/Scripts/RidersGame/VehicleSystems/State Machine/MovementStateMachine.cs
@@ -1,5 +1,3 @@
-using UnityEngine;
-
 namespace RidersRuntime.VehicleSystem
 {
     public class MovementStateMachine
